requirements.txt
Django==1.9.9
django-filer==1.2.4
django-jinja==2.2.1
django-mptt==0.8.5
django-nested-inline==0.3.6
django-polymorphic==0.8.1
easy-thumbnails==2.3
Jinja2==2.8
Pillow==3.3.0
psycopg2==2.6.2
python-social-auth==0.2.21
requests==2.11.*
uWSGI==2.0.*
django-import-export==0.4.*
django-admin-sortable2==0.6.*
openpyxl
